refactor(introspect): split rethink.rs (2132 lines) into a rethink/ module
Pure mechanical split — no behavior change, all 225 tests preserved and green. The
2132-line cmd/rethink.rs becomes a directory module with focused submodules, each
well under the 500-line review line:

  mod.rs       37   declarations + re-exports
  scope.rs     78   project scoping
  model.rs    135   WayEvent/ActiveWay/Frame/View/Player/TermGuard
  keys.rs     236   timeline + drill-down key handlers, anchor navigation
  sessions.rs 297   session listing + picker
  run.rs      338   run/run_live/tui_loop/refresh (entry points + loops)
  layout.rs   345   compose_screen/render_frame/status bar/header/justify
  frames.rs   374   build_frames + windowing + token timeline + event loading
  drilldown.rs 429  why-index + render_why + detail

External refs (crate::cmd::rethink::{...}) resolve via re-exports in mod.rs;
intra-module items lifted to pub(super); all cfg(feature="tui") gates preserved.
Addresses the code-quality file-length flag (was task #21).
